systemcontrol: Modified hist param struct define[1/2]
PD# 173911

Modified hist param struct define.

(cherry picked from commit d282f6b1f13858230cdbcfcfdd31ad143238f402)

+++ services/systemcontrol/PQ/include/ve.h
@@ -132,7 +132,7 @@ typedef struct ve_dnlp_s {
     enum ve_dnlp_ext_e white;
 } ve_dnlp_t;
 typedef struct ve_hist_s {
-	unsigned long long sum;
+    unsigned int sum;
     int width;
     int height;
     int ave;
